Converted SelectInput Component to Use React Hooks
Tha SelectInput component now uses React Hooks. When the screen size changes, the current selection title changes to the max length string for a second and then switches back to the correct selection. This always happened, but is visible to the user now that the component uses hooks. The underlying idea behind the measurement should be changed.